update the folder

+++ add.py
@@ -1,4 +1,6 @@
 a=int(input("enter 1st no :"))
-b=int(input("enter 2nd no :"))
+b=float(input("enter 2nd no :"))
 
 print(a+b)
+
+
